refactor: renderTooltip
