+++ packages/superset-ui-preset-chart-xy/src/Line/renderTooltip.tsx
@@ -1,0 +1,55 @@
+/* eslint-disable no-magic-numbers */
+
+import React from 'react';
+import { chartTheme } from '@data-ui/theme';
+import TooltipFrame from '../components/tooltip/TooltipFrame';
+import TooltipTable from '../components/tooltip/TooltipTable';
+import { TooltipInput } from './Line';
+
+const MARK_STYLE = { marginRight: 4 };
+
+export default function renderTooltip({
+  allSeries,
+  datum,
+  encoder,
+  series = {},
+  theme = chartTheme,
+}: TooltipInput) {
+  return (
+    <TooltipFrame>
+      <>
+        <div style={{ fontFamily: theme.labelStyles.fontFamily }}>
+          <strong>{encoder.channels.x.formatValue(datum.x)}</strong>
+        </div>
+        <br />
+        {series && (
+          <TooltipTable
+            data={allSeries
+              .filter(({ key }) => series[key])
+              .concat()
+              .sort((a, b) => series[b.key].y - series[a.key].y)
+              .map(({ key, stroke, strokeDasharray }) => ({
+                key,
+                keyColumn: (
+                  <>
+                    <svg width="12" height="8" style={MARK_STYLE}>
+                      <line
+                        x2="12"
+                        y1="3"
+                        y2="3"
+                        stroke={stroke}
+                        strokeWidth="2"
+                        strokeDasharray={strokeDasharray}
+                      />
+                    </svg>
+                    {series[key] === datum ? <b>{key}</b> : key}
+                  </>
+                ),
+                valueColumn: encoder.channels.y.formatValue(series[key].y),
+              }))}
+          />
+        )}
+      </>
+    </TooltipFrame>
+  );
+}
